:building_construction: Use double values in post render event

+++ src/main/java/net/inceptioncloud/dragonfly/event/client/PostRenderEvent.kt
@@ -10,8 +10,8 @@ import net.minecraft.client.renderer.EntityRenderer
  * has finished, this event will be fired.
  */
 data class PostRenderEvent(
-    val scaledWidth: Int,
-    val scaledHeight: Int,
-    val scaledMouseX: Int,
-    val scaledMouseY: Int
+    val scaledWidth: Double,
+    val scaledHeight: Double,
+    val scaledMouseX: Double,
+    val scaledMouseY: Double
 )
